+++ api/app/Repositories/UserRepository.php
@@ -82,7 +82,7 @@ class UserRepository extends Controller
                 $limit = $params['limit'];
             }
             
-            $users = SecurityUsers::with('identityType', 'nationalities', 'identities');
+            $users = SecurityUsers::with('identityType', 'nationalities', 'identities', 'institutionStaff', 'institutionStaff.staffPositionGrade');
             if(isset($params['order'])){
                 $orderBy = $params['order_by']??"ASC";
                 $col = $params['order'];
@@ -116,7 +116,9 @@ class UserRepository extends Controller
                     'institutionStudent.studentStatus',
                     'identities',
                     'nationality',
-                    'identityType'
+                    'identityType',
+                    'institutionStaff',
+                    'institutionStaff.staffPositionGrade'
                 )
                     ->where('id', $userId)
                     ->get();